Cover area with tests
